pyproject.toml
[project]
name = "pleskal"
version = "0.1.0"
description = "A local-first, crowd-sourced web application for discovering and sharing dance events in Copenhagen."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "psycopg[binary]>=3.2",
    "django-markdownx>=4.0",
    "nh3>=0.2",
    "django-storages[boto3]>=1.14",
    "django-axes>=7.0",
    "django-environ>=0.12",
    "pillow>=12.1.1",
    "gunicorn>=23.0",
    "sentry-sdk[django]>=2.0",
    "whitenoise>=6.8",
    "icalendar>=6.0",
    "pytest-xdist>=3.8.0",
    "zxcvbn>=4.4",
    "requests>=2.32.5",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "django-anymail[resend]>=10.0",
    "markdownify>=1.2.2",
    "cryptography>=46.0.5",
]

[dependency-groups]
dev = [
    "ruff>=0.9",
    "pytest>=8.0",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "factory-boy>=3.3",
    "django-debug-toolbar>=5.0",
    "ty>=0.0.23",
]

[tool.ruff]
target-version = "py313"
line-length = 88
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors: style and syntax violations
    "F",   # Pyflakes: undefined names, unused imports, and other logical errors
    "I",   # isort: enforce consistent import ordering
    "UP",  # pyupgrade: modernize Python syntax for newer language versions
    "B",   # flake8-bugbear: catch likely bugs and opinionated design problems
    "SIM", # flake8-simplify: suggest simplifications to existing code constructs
    "S",   # flake8-bandit: detect common security vulnerabilities
]
ignore = [
    "E501", # line too long: not enforced, handled by the formatter
]

[tool.ruff.lint.per-file-ignores]
"*/tests/*" = [
    "S101", # assert: assert statements are idiomatic and expected in tests
    "S106", # hardcoded-password-func-arg: literal passwords are routine in test fixtures
    "S314", # suspicious-xml-element-tree-usage: parsing trusted test response data, not user input
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py"]
addopts = "--reuse-db -n 8"

[tool.coverage.run]
source = ["."]
omit = ["*/migrations/*", "*/tests/*", "manage.py", "config/*"]
relative_files = true
